Support no autocomplete delimiters
With no autocomplete delimiters, all of the characters to the left of
the cursor are compared for a match. This is useful in situations where
it is not important for the completion to occur in the middle of the
text, and it is perhaps more important that delimiter characters are
allowed in the phrase(s) to autocomplete themselves.

One example case this helps is where the autocomplete phrase may contain a
space (eg. multi-part personal names); if the delimiter set also
includes a space then all name-parts before the space may be neglected
in the match. This is particularly apparent when backspacing and
autocompleting multiple times.

Tests added